Reworked build system
Some larger rework on the build system with improved UI
and a remote filestorage as a new docker container.
This allows running the build/exampels on a arbitary machine and
push the results, including e.g. h5 file to www.mbsime-env.de
This allows to run the build/examples on e.g. external hosters.