Prepare for switching to another CentOS mirror since the official mirror ended since 2024-07-01
This commits changes the mirror only for webapprunImage for now since
this is needed always.

At all other places just a comment is added about how to change the
mirror. If a rebuild of such a part is needed this comment must be
removed to enable "yum" again at this place.

The the long term we should switch to another distro!